♻️ refactor: randomLotto 생성 기능 분리
