+++ src/step2-index.js
@@ -1,9 +1,8 @@
 import generateAnswerLotto from "./domain/AnswerLottoPack.js";
-import LottoMachine from "./domain/LottoMachine.js";
 import profitCalculator from "./domain/profitCalculator.js";
-import validatePurchaseAmount from "./domain/validation/validatePurchaseAmount.js";
 import validateWinningNumbers from "./domain/validation/validateWinningNumbers.js";
 import validateBonusNumber from "./domain/validation/validateBonusNumber.js";
+import { randomLottos } from "./web/randomLottos/randomLottos.js";
 /**
  * step 2의 시작점이 되는 파일입니다.
  * 노드 환경에서 사용하는 readline 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
@@ -16,47 +15,18 @@ const resultSection = document.querySelector(".result_section");
 const winningSection = document.querySelector(".winning_number_section");
 const restartButton = document.querySelector(".restart");
 
-purchaseButton.addEventListener("click", () => {
-  const { purchaseAmount, lottoPack } = lottoStart();
-  resultSection.style.display = "block";
-  winningSection.style.display = "block";
-  resultButton.addEventListener("click", () => {
-    checkStart(purchaseAmount, lottoPack);
-
-    restartButton.addEventListener("click", restart);
-    closeButton.addEventListener("click", () => {
-      modal.style.display = "none";
-    });
-  });
+document.addEventListener("DOMContentLoaded", () => {
+  purchaseButton.addEventListener("click", randomLottos);
 });
 
-const lottoStart = () => {
-  try {
-    // 구입 금액 입력 도메인 연결
-    const purchaseAmount = document.querySelector(".purchase_input").value;
-    validatePurchaseAmount(purchaseAmount);
-
-    const { count, lottoPack } = LottoMachine(purchaseAmount);
-    const resultText = document.querySelector(".result_text");
-    resultText.textContent = `총 ${count}개를 구매하였습니다.`;
-
-    // 랜덤 로또 번호 출력 도메인 연결
-    const lottoPackContainer = document.querySelector(".lotto_pack");
-    lottoPackContainer.innerHTML = "";
-    lottoPack.lottos.forEach((lotto) => {
-      const lottoElement = document.createElement("div");
-      lottoElement.classList.add("lotto");
-      lottoElement.innerHTML = `
-      <img src="./public/ticket.png" />
-      <p class="random_number">${lotto.lottoNumbers.join(", ")}</p>
-    `;
-      lottoPackContainer.appendChild(lottoElement);
-    });
-    return { purchaseAmount, lottoPack };
-  } catch (error) {
-    window.alert(error.message);
-    document.querySelector(".purchase_input").value = "";
-  }
+export const playLotto = (purchaseAmount, lottoPack) => {
+  resultButton.addEventListener("click", () => {
+    checkStart(purchaseAmount, lottoPack);
+  });
+  restartButton.addEventListener("click", restart);
+  closeButton.addEventListener("click", () => {
+    modal.style.display = "none";
+  });
 };
 
 const checkStart = (purchaseAmount, lottoPack) => {
